+++ gemvs_real.cu
@@ -54,7 +54,7 @@ __global__ void minus_transAx_plusB_scale_real(real_t scale, int32_t M, int32_t 
   minus_a_fma_real fma_func;
 
   for (int32_t row = blockIdx.x; row < M; row += GRID_BLOCKS) {
-    real_const_ptr A_i = &A[row * lda];
+    real_const_ptr A_i = &A[uint64_t(row) * uint64_t(lda)];
 
     block_load.Load(A_i, threadA, N1, real_t());
     block_load.Load(A, threadX, N1, real_t());
@@ -89,7 +89,7 @@ __global__ void minus_transAx_plusB_scale_real(real_t scale, int32_t M, int32_t 
 
       real_t res = scal_func(block_res, B[row], scale);
       B[row] = res;
-      B[row * lda] = res;
+      B[uint64_t(row) * uint64_t(lda)] = res;
     }
   }
 }
